# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This describes the REST API for Japonic auction apps",
      "version": "1.0.0",
      "title": "Japonic API"
    },
    "host": "http://192.168.0.73:3000/api",
    "schemes": [
      "http"
    ],
    "security": [
      {
        "basicAuth": []
      }
    ],
    "paths": {
      "/companies?limit={limit}&page_number={page_number}": {
        "get": {
          "tags": [
            "The Endpoint to get companies"
          ],
          "summary": "Endpoints to get companies",
          "description": "Endpoints to get companies",
          "operationId": "wakeup",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "text/json"
          ],
          "parameters": [
            {
              "name": "limit",
              "in": "path",
              "description": "The identifier of the limit amount of companies to load per a request",
              "required": true,
              "type": "number"
            },
            {
              "name": "page_number",
              "in": "path",
              "description": "Which page to get companies in",
              "required": true,
              "type": "number"
            }
          ],
          "responses": {
            "200": {
                "description": "A list of companies.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                            "$ref": "#/components/schemas/pet"
                            }
                        }
                    }
                }
            },
            "422": {
              "description": "We were unable to move the tablet to image capture mode"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      },
      "/models?limit={limit}&page_number={page_number}" : {
        "get": {
          "tags": [
            "The Endpoint to get companies"
          ],
          "summary": "Endpoints to get companies",
          "description": "Endpoints to get companies",
          "operationId": "wakeup",
          "consumes": [
            "multipart/form-data"
          ],
          "produces": [
            "text/json"
          ],
          "parameters": [
            {
              "name": "limit",
              "in": "path",
              "description": "The identifier of the limit amount of companies to load per a request",
              "required": true,
              "type": "number"
            },
            {
              "name": "page_number",
              "in": "path",
              "description": "Which page to get companies in",
              "required": true,
              "type": "number"
            }
          ],
          "responses": {
            "200": {
                "description": "A list of companies.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                            "$ref": "#/components/schemas/pet"
                            }
                        }
                    }
                }
            },
            "422": {
              "description": "We were unable to move the tablet to image capture mode"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "API Certificate"
      }
    },
...
  }